refactor category
